// src/app/manage/vehicle-manage/model/upkeep-record-req.model.ts
export class UpkeepRecordReq {
    id: number;
    acceptancePerson: string = null;
    chargePerson: string = null;
    planUpkeepTime: number = null;
    plateNumber: string = null;
    upkeepCompany: string = null;
    upkeepContent: string = null;
    upkeepCost: number = null;
    upkeepMileage: number = null;
    upkeepPerson: string = null;
    upkeepTime: number = null;
    upkeepTimeCost: number = null;
    vehicleId: number = null;
}